moved tcp resize communication into seperate library
